+++ test/ILP/ilp-credentials-test.web.ts
@@ -1,11 +1,11 @@
 import { assert } from 'chai'
-import IlpCredentials from '../../src/ILP/ilp-credentials.web'
+import IlpWebCredentials from '../../src/ILP/auth/ilp-credentials.web'
 
 describe('IlpCredentials Web', function(): void {
   it('Build - Undefined token', function(): void {
     // GIVEN no bearer token
     // WHEN IlpCredentials are built without a bearer token
-    const credentials = IlpCredentials.build()
+    const credentials = IlpWebCredentials.build()
 
     // THEN credentials should be undefined
     assert(!credentials)
@@ -16,7 +16,7 @@ describe('IlpCredentials Web', function(): void {
     const bearerToken = 'password'
 
     // WHEN IlpCredentials are built
-    const credentials = IlpCredentials.build(bearerToken)
+    const credentials = IlpWebCredentials.build(bearerToken)
 
     // THEN "Bearer " is added to token
     assert.equal(
@@ -30,7 +30,7 @@ describe('IlpCredentials Web', function(): void {
     const bearerToken = 'Bearer password'
 
     // WHEN IlpCredentials are built
-    const credentials = IlpCredentials.build(bearerToken)
+    const credentials = IlpWebCredentials.build(bearerToken)
 
     // THEN bearer token is unchanged
     assert.equal(credentials && credentials.Authorization, bearerToken)
